+++ app/src/main/java/com/opentasker/ui/screens/ActiveAutomationUi.kt
@@ -919,6 +919,7 @@ fun ActiveAutomationUi(
                 contentPadding = innerPadding,
                 onRefresh = viewModel::refreshDiagnostics,
                 onShare = viewModel::shareDiagnosticReport,
+                onCopy = viewModel::copyDiagnosticReport,
             )
         }
             }